[functions] Move listTextureSlots, listTextureChannels to /functions package
